[#67]refactor: AuthService 내부 요청을 route handler 기반으로 분리
